+++ src/pages/Profile/components/index.ts
@@ -1,2 +1,1 @@
-export { default as Header } from './Header';
 export { default as Body } from './Body';
