+++ src/pages/Profile/styles.module.scss
@@ -1,6 +1,102 @@
 @import 'src/styles/helpers';
 
-.profile{
+.profile {
   @include container;
   max-width: 1440px;
+  .header {
+    @include flex(row, flex-start, flex-start);
+    margin-top: space(65);
+
+    .right {
+      @include flex(row, flex-start, space-between);
+      width: 100%;
+    }
+    .info {
+      margin-left: space(80);
+    }
+
+    .following {
+      @include flex(row, flex-start, flex-start);
+      .followers,
+      .follows {
+        height: 32px;
+        min-width: 120px;
+        border: 1px solid;
+        box-shadow: none;
+      }
+      .followers {
+        border-color: $accent;
+        background-color: $light-100;
+      }
+      .follows {
+        border-color: $yellow-500;
+        background-color: $yellow-50;
+        &:hover {
+          background-color: $yellow-50;
+        }
+      }
+      .followers + .follows {
+        margin-left: space(10);
+      }
+    }
+
+    .balances {
+      @include flex(row, flex-start, flex-start);
+      margin-top: space(16);
+      margin-bottom: space(24);
+      .balanceItem {
+        @include flex(row, center, flex-start);
+      }
+      .balanceItem + .balanceItem {
+        margin-left: space(27);
+      }
+      .balanceText {
+        font-size: 20px;
+        line-height: 34px;
+        margin-left: space(8);
+      }
+    }
+
+    .edit {
+      margin-top: space(16);
+    }
+  }
+  
+    .socials {
+      background-color: $light-white;
+      padding: space(24);
+      border-radius: 14px;
+      &Item {
+        @include flex(row, center, flex-start);
+      }
+      &Icon {
+        margin-right: space(11);
+        width: 28px;
+        height: 28px;
+        border-radius: 50%;
+        background-color: $light-100;
+        @include flex(row, center, center);
+      }
+      .socialsItem + .socialsItem {
+        margin-top: space(16);
+      }
+    }
+}
+@include large-tablet {
+  .profile {
+    .header {
+      .info {
+        margin-left: space(32);
+      }
+    }
+  }
+}
+@include middle-tablet{
+  .profile{
+    .socials{
+      background-color: transparent;
+      padding: 0;
+      margin-top: 93px;
+    }
+  }
 }
